updated register and profile pages
fixed spelling error in register page. Updated some of the styling for the profile page to show a better looking profile page. Updated the way the user edits their profile with a button that then shows the inputs rather than them always being there. Added a submit and go back button if the user is editing their profile and changed some other styling for user image and like-dislike chart. Still more work to be done with styling though.