Updated profile, login and signup pages
Profile page now has user dropdown menu to change size along and updated some styling

the signup page styling changed and added the drop downs for breed and size to that too. The button to go back to the login page styling changed.

the login page styling changed for the button to change to the register page